examples/fastapi/pyproject.toml
[project]
name = "event-driven-fastapi-service"
version = "0.1.0"
description = "Event-driven system with acknowledgment tracking and extensible event classes"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.9.0",
    "rx-events>=0.1.0",
    "fastapi>=0.133.1",
    "pydantic-settings>=2.13.1",
    "reactivex>=4.1.0",
    "uvicorn>=0.41.0",
]
readme = "README.md"
license = {text = "MIT"}

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
    "httpx>=0.24.0",
]


[tool.hatchling.build.targets.wheel]
only-include = ["src"]